+++ docs/CHANGELOG.rst
@@ -19,6 +19,7 @@ Changed
 - Make changes to input fields in ``workflow-cutnrun-beta``
 - Reword documentation of ``macs2-callpeak``, ``macs2-batch`` and
   ``qc-prepeak``
+- Add input option to exclude flags in ``samtools-bedcov`` process
 
 Fixed
 -----